Better EULA configuring

+++ src/main/groovy/ru/endlesscode/bukkitgradle/server/PrepareServer.groovy
@@ -31,9 +31,14 @@ class PrepareServer extends DefaultTask {
 
     void resolveEula() {
         Path eulaFile = serverDir.resolve("eula.txt")
+        if (!Files.exists(eulaFile)) {
+            Files.createFile(eulaFile)
+        }
 
-        boolean eula = this.run.eula
-        eulaFile.text = "eula=$eula"
+        Properties properties = new Properties()
+        properties.load(eulaFile.newReader())
+        properties.setProperty("eula", "${this.run.eula}")
+        properties.store(eulaFile.newWriter(), "By changing the setting below to TRUE you are indicating your agreement to our EULA (https://account.mojang.com/documents/minecraft_eula).")
     }
 
     void resolveOnlineMode() {
@@ -51,9 +56,12 @@ class PrepareServer extends DefaultTask {
     void copyPluginToServerDir() {
         String pluginName = "${project.bukkit.meta.name}.jar"
         Path jar = project.jar.archivePath.toPath()
+        if (!Files.exists(jar)) {
+            return
+        }
+        
         Path pluginsDir = serverDir.resolve("plugins")
         Files.createDirectories(pluginsDir)
-
         Files.copy(jar, pluginsDir.resolve(pluginName), StandardCopyOption.REPLACE_EXISTING)
     }
 }
